# takeout/src/main/java/hu/oe/takeout/takeout/generated/rest/api/DefaultApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.oe.takeout.takeout.generated.rest.api;

import hu.oe.takeout.takeout.generated.rest.model.CategoriesIdGet404Response;
import hu.oe.takeout.takeout.generated.rest.model.CategoryRequest;
import hu.oe.takeout.takeout.generated.rest.model.CategoryResponse;
import hu.oe.takeout.takeout.generated.rest.model.IdModel;
import hu.oe.takeout.takeout.generated.rest.model.TakeoutIdGet404Response;
import hu.oe.takeout.takeout.generated.rest.model.TakeoutRequest;
import hu.oe.takeout.takeout.generated.rest.model.TakeoutResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "Default", description = "the Default API")
public interface DefaultApi {

    /**
     * GET /categories : Ételcsoportok listázása
     * Lekérdezi az összes ételcsoportot.
     *
     * @return Az ételcsoportok sikeresen lekérdezve. (status code 200)
     */
    @Operation(
        operationId = "categoriesGet",
        summary = "Ételcsoportok listázása",
        description = "Lekérdezi az összes ételcsoportot.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az ételcsoportok sikeresen lekérdezve.", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = CategoryResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/categories",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<CategoryResponse>> categoriesGet(
        
    );


    /**
     * DELETE /categories/{id} : Ételcsoport törlése
     * Törli az ételcsoportot az ID alapján.
     *
     * @param id Az ételcsoport egyedi azonosítója. (required)
     * @return Az ételcsoport sikeresen törölve (status code 204)
     *         or Az ételcsoport nem található (status code 404)
     */
    @Operation(
        operationId = "categoriesIdDelete",
        summary = "Ételcsoport törlése",
        description = "Törli az ételcsoportot az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Az ételcsoport sikeresen törölve"),
            @ApiResponse(responseCode = "404", description = "Az ételcsoport nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoriesIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/categories/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> categoriesIdDelete(
        @Parameter(name = "id", description = "Az ételcsoport egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /categories/{id} : Ételcsoport lekérdezése
     * Lekérdezi egy adott ételcsoport az ID alapján.
     *
     * @param id Az ételcsoport egyedi azonosítója. (required)
     * @return Az ételcsoport sikeresen lekérdezve (status code 200)
     *         or Az ételcsoport nem található (status code 404)
     */
    @Operation(
        operationId = "categoriesIdGet",
        summary = "Ételcsoport lekérdezése",
        description = "Lekérdezi egy adott ételcsoport az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az ételcsoport sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Az ételcsoport nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoriesIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/categories/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<CategoryResponse> categoriesIdGet(
        @Parameter(name = "id", description = "Az ételcsoport egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * PUT /categories/{id} : Ételcsoport módosítása
     * Módosítja az ételcsoport nevét az ID alapján.
     *
     * @param id Az ételcsoport egyedi azonosítója. (required)
     * @param categoryRequest  (required)
     * @return Az ételcsoport sikeresen módosítva (status code 200)
     *         or Az ételcsoport nem található (status code 404)
     */
    @Operation(
        operationId = "categoriesIdPut",
        summary = "Ételcsoport módosítása",
        description = "Módosítja az ételcsoport nevét az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az ételcsoport sikeresen módosítva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Az ételcsoport nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoriesIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/categories/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<IdModel> categoriesIdPut(
        @Parameter(name = "id", description = "Az ételcsoport egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "CategoryRequest", description = "", required = true) @Valid @RequestBody CategoryRequest categoryRequest
    );


    /**
     * POST /categories : Új ételcsoport hozzáadása
     * Új ételcsoportot ad hozzá a nyilvántartáshoz.
     *
     * @param categoryRequest  (required)
     * @return Az ételcsoport sikeresen hozzáadva (status code 201)
     *         or A csoport már létezik (status code 404)
     */
    @Operation(
        operationId = "categoriesPost",
        summary = "Új ételcsoport hozzáadása",
        description = "Új ételcsoportot ad hozzá a nyilvántartáshoz.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Az ételcsoport sikeresen hozzáadva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "A csoport már létezik", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/categories",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<IdModel> categoriesPost(
        @Parameter(name = "CategoryRequest", description = "", required = true) @Valid @RequestBody CategoryRequest categoryRequest
    );


    /**
     * GET /takeout : Ételek listázása
     * Lekérdezi az összes ételt.
     *
     * @return Az ételek sikeresen lekérdezve (status code 200)
     */
    @Operation(
        operationId = "takeoutGet",
        summary = "Ételek listázása",
        description = "Lekérdezi az összes ételt.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az ételek sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = TakeoutResponse.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/takeout",
        produces = { "application/json" }
    )
    
    ResponseEntity<List<TakeoutResponse>> takeoutGet(
        
    );


    /**
     * DELETE /takeout/{id} : Étel törlése
     * Törli az ételt az ID alapján.
     *
     * @param id Az étel egyedi azonosítója. (required)
     * @return Az étel sikeresen törölve (status code 204)
     *         or Az étel nem található (status code 404)
     */
    @Operation(
        operationId = "takeoutIdDelete",
        summary = "Étel törlése",
        description = "Törli az ételt az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "204", description = "Az étel sikeresen törölve"),
            @ApiResponse(responseCode = "404", description = "Az étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TakeoutIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/takeout/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<Void> takeoutIdDelete(
        @Parameter(name = "id", description = "Az étel egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * GET /takeout/{id} : Étel lekérdezése
     * Lekérdezi egy adott ételt az ID alapján.
     *
     * @param id Az étel egyedi azonosítója. (required)
     * @return Az étel sikeresen lekérdezve (status code 200)
     *         or Az étel nem található (status code 404)
     */
    @Operation(
        operationId = "takeoutIdGet",
        summary = "Étel lekérdezése",
        description = "Lekérdezi egy adott ételt az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az étel sikeresen lekérdezve", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TakeoutResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Az étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TakeoutIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/takeout/{id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<TakeoutResponse> takeoutIdGet(
        @Parameter(name = "id", description = "Az étel egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id
    );


    /**
     * PUT /takeout/{id} : Étel módosítása
     * Módosítja az étel nevét az ID alapján.
     *
     * @param id Az étel egyedi azonosítója. (required)
     * @param takeoutRequest  (required)
     * @return Az étel sikeresen módosítva (status code 200)
     *         or Az étel nem található (status code 404)
     */
    @Operation(
        operationId = "takeoutIdPut",
        summary = "Étel módosítása",
        description = "Módosítja az étel nevét az ID alapján.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Az étel sikeresen módosítva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Az étel nem található", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = TakeoutIdGet404Response.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/takeout/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<IdModel> takeoutIdPut(
        @Parameter(name = "id", description = "Az étel egyedi azonosítója.", required = true, in = ParameterIn.PATH) @PathVariable("id") String id,
        @Parameter(name = "TakeoutRequest", description = "", required = true) @Valid @RequestBody TakeoutRequest takeoutRequest
    );


    /**
     * POST /takeout : Új étel hozzáadása
     * Új ételt ad hozzá a nyilvántartáshoz.
     *
     * @param takeoutRequest  (required)
     * @return Az étel sikeresen hozzáadva (status code 201)
     *         or Az étel már létezik (status code 404)
     */
    @Operation(
        operationId = "takeoutPost",
        summary = "Új étel hozzáadása",
        description = "Új ételt ad hozzá a nyilvántartáshoz.",
        responses = {
            @ApiResponse(responseCode = "201", description = "Az étel sikeresen hozzáadva", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = IdModel.class))
            }),
            @ApiResponse(responseCode = "404", description = "Az étel már létezik", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/takeout",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<IdModel> takeoutPost(
        @Parameter(name = "TakeoutRequest", description = "", required = true) @Valid @RequestBody TakeoutRequest takeoutRequest
    );

}