# takeout/src/main/resources/openapi/takeout.yaml
openapi: 3.0.0
info:
  title: Étel Nyilvántartás API
  description: "API az ételek nyilvántartására, ahol minden ételhez egyedi azonosító\
    t és nevet rendelünk."
  version: 1.0.0
servers:
- url: http://localhost:8080/api
  description: Helyi fejlesztői szerver
paths:
  /categories:
    get:
      summary: Ételcsoportok listázása
      description: Lekérdezi az összes ételcsoportot.
      responses:
        "200":
          description: Az ételcsoportok sikeresen lekérdezve.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
                example:
                  - id: 1
                    name: Tészta
                  - id: 2
                    name: Leves
    post:
      summary: Új ételcsoport hozzáadása
      description: Új ételcsoportot ad hozzá a nyilvántartáshoz.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryRequest"
      responses:
        "201":
          description: Az ételcsoport sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdModel"
              example:
                id: 4
                name: Desszert
        "404":
          description: "A csoport már létezik"
          content:
            application/json:
              schema:
                type: object
                schema:
                  $ref: "#/components/schemas/ApiError"
                example:
                  path: "/categories"
                  message: "error.validation"
                  errors:
                    - "error.categories.name.exists"
  /categories/{id}:
      get:
        summary: Ételcsoport lekérdezése
        description: Lekérdezi egy adott ételcsoport az ID alapján.
        parameters:
          - in: path
            name: id
            required: true
            schema:
              type: string
            description: Az ételcsoport egyedi azonosítója.
        responses:
          "200":
            description: Az ételcsoport sikeresen lekérdezve
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/CategoryResponse"
                example:
                  id: 1
                  name: Tészta
          "404":
            description: Az ételcsoport nem található
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: Ételcsoport nem található
      put:
        summary: Ételcsoport módosítása
        description: Módosítja az ételcsoport nevét az ID alapján.
        parameters:
          - in: path
            name: id
            required: true
            schema:
              type: string
            description: Az ételcsoport egyedi azonosítója.
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryRequest"
        responses:
          "200":
            description: Az ételcsoport sikeresen módosítva
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/IdModel"
                example:
                  id: 1
                  name: Gomba
          "404":
            description: Az ételcsoport nem található
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: Ételcsoport nem található
      delete:
        summary: Ételcsoport törlése
        description: Törli az ételcsoportot az ID alapján.
        parameters:
          - in: path
            name: id
            required: true
            schema:
              type: string
            description: Az ételcsoport egyedi azonosítója.
        responses:
          "204":
            description: Az ételcsoport sikeresen törölve
          "404":
            description: Az ételcsoport nem található
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    error:
                      type: string
                      example: Ételcsoport nem található

  /takeout:
    get:
      summary: Ételek listázása
      description: Lekérdezi az összes ételt.
      responses:
        "200":
          description: Az ételek sikeresen lekérdezve
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TakeoutResponse"
              example:
              - id: 1
                name: Hamburger
                price: 1500
                categoryId: 1
              - id: 2
                name: Pizza
                price: 1500
                categoryId: 2
              - id: 3
                name: Rántott hús
                price: 1500
                categoryId: 3
    post:
      summary: Új étel hozzáadása
      description: Új ételt ad hozzá a nyilvántartáshoz.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TakeoutRequest"
      responses:
        "201":
          description: Az étel sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdModel"
              example:
                id: 4
                name: Rántotta
                price: 550
                categoryId: 1
        "404":
          description: Az étel már létezik
          content:
            application/json:
              schema:
                type: object
                schema:
                  $ref: "#/components/schemas/ApiError"
                example:
                  path: "/takeout"
                  message: "error.validation"
                  errors:
                    - "error.takeout.name.exists"
  /takeout/{id}:
    get:
      summary: Étel lekérdezése
      description: Lekérdezi egy adott ételt az ID alapján.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Az étel egyedi azonosítója.
      responses:
        "200":
          description: Az étel sikeresen lekérdezve
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TakeoutResponse"
              example:
                id: 1
                name: Hamburger
                price: 1500
                categoryId: 1
        "404":
          description: Az étel nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Étel nem található
    put:
      summary: Étel módosítása
      description: Módosítja az étel nevét az ID alapján.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Az étel egyedi azonosítója.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TakeoutRequest"
      responses:
        "200":
          description: Az étel sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdModel"
              example:
                id: 1
                name: Muffin
                price: 1500
                categoryId: 1
        "404":
          description: Az étel nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Étel nem található
    delete:
      summary: Étel törlése
      description: Törli az ételt az ID alapján.
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
        description: Az étel egyedi azonosítója.
      responses:
        "204":
          description: Az étel sikeresen törölve
        "404":
          description: Az étel nem található
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Étel nem található
components:
  schemas:
    IdModel:
      type: object
      properties:
        id:
          type: string

    TakeoutRequest:
      type: object
      required:
      - name
      - price
      - id
      properties:
        name:
          type: string
          description: Étel neve
          x-field-extra-annotation: '@hu.oe.takeout.validation.NameExists(message="error.takeout.name.exists")'
        price:
          type: integer
          description: Az étel ára.
        categoryId:
          type: string
          description: Az ételhez tartozó csoport.
        description:
          type: string
          description: Leírás
    TakeoutResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/IdModel'
        - $ref: '#/components/schemas/TakeoutRequest'

    CategoryRequest:
      type: object
      required:
        - id
        - name
      properties:
        name:
          type: string
          description: Kategória neve
          x-field-extra-annotation: '@hu.oe.takeout.validation.NameExists(message="error.category.name.exists")'
        description:
          type: string
          description: Leírás
    CategoryResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/IdModel'
        - $ref: '#/components/schemas/CategoryRequest'

    ApiError:
      type: object
      properties:
        path:
          type: string
        message:
          type: string
        errors:
          type: array
          items: string